pyproject.toml
[tool.poetry]
name = "panel_client"
version = "0.1.0"
description = ""
authors = ["Evan Nishi <evan@evannishi.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.19.2"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
